Refactor FlutterShellArgs
