+++ engine/src/flutter/shell/platform/android/io/flutter/embedding/android/FlutterActivity.java
@@ -49,7 +49,7 @@ import androidx.lifecycle.LifecycleRegistry;
 import io.flutter.Log;
 import io.flutter.embedding.android.FlutterActivityLaunchConfigs.BackgroundMode;
 import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.embedding.engine.FlutterShellArgsIntentUtils;
+import io.flutter.embedding.engine.FlutterShellArgs;
 import io.flutter.embedding.engine.plugins.activity.ActivityControlSurface;
 import io.flutter.embedding.engine.plugins.util.GeneratedPluginRegister;
 import io.flutter.plugin.platform.PlatformPlugin;
@@ -1042,8 +1042,8 @@ public class FlutterActivity extends Activity
    */
   @NonNull
   @Override
-  public String[] getFlutterShellArgs() {
-    return FlutterShellArgsIntentUtils.getFlutterShellCommandLineArgs(getIntent());
+  public FlutterShellArgs getFlutterShellArgs() {
+    return FlutterShellArgs.fromIntent(getIntent());
   }
 
   /**
