+++ engine/src/flutter/shell/platform/android/io/flutter/embedding/android/FlutterActivityAndFragmentDelegate.java
@@ -28,6 +28,7 @@ import io.flutter.Build.API_LEVELS;
 import io.flutter.FlutterInjector;
 import io.flutter.Log;
 import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.embedding.engine.FlutterEngineFlags;
 import io.flutter.embedding.engine.FlutterEngineCache;
 import io.flutter.embedding.engine.FlutterEngineGroup;
 import io.flutter.embedding.engine.FlutterEngineGroupCache;
@@ -333,11 +334,9 @@ import java.util.Set;
             + " this FlutterFragment.");
 
     warnIfEngineFlagsSetViaIntent(host.getActivity().getIntent());
-    String[] flutterShellArgs =
-        host.getFlutterShellArgs() == null ? new String[0] : host.getFlutterShellArgs();
     FlutterEngineGroup group =
         engineGroup == null
-            ? new FlutterEngineGroup(host.getContext(), flutterShellArgs)
+            ? new FlutterEngineGroup(host.getContext(), host.getFlutterShellArgs().toArray())
             : engineGroup;
     flutterEngine =
         group.createAndRunEngine(
@@ -360,13 +359,13 @@ import java.util.Set;
     Set<String> extrasKeys = extras.keySet();
 
     for (String extrasKey : extrasKeys) {
-      FlutterShellArgs.Flag flag = FlutterShellArgs.getFlagFromIntentKey(extrasKey);
+      FlutterEngineFlags.Flag flag = FlutterEngineFlags.getFlagFromIntentKey(extrasKey);
       if (flag != null) {
         Log.w(
             TAG,
             "Support for setting engine flags on Android via Intent will soon be dropped; see https://github.com/flutter/flutter/issues/180686 for more information on this breaking change. To migrate, set "
                 + flag.commandLineArgument
-                + " on the command line or see https://github.com/flutter/flutter/blob/main/docs/engine/Android-Flutter-Shell-Arguments.md for alternative methods.");
+                + " on the command line or see https://github.com/flutter/flutter/blob/main/docs/engine/Android-Flutter-Engine-Flags.md for alternative methods.");
         break;
       }
     }
@@ -1119,7 +1118,7 @@ import java.util.Set;
     Lifecycle getLifecycle();
 
     @NonNull
-    String[] getFlutterShellArgs();
+    FlutterShellArgs getFlutterShellArgs();
 
     /**
      * Returns the ID of a statically cached {@link io.flutter.embedding.engine.FlutterEngine} to
