+++ engine/src/flutter/shell/platform/android/io/flutter/embedding/android/FlutterFragment.java
@@ -27,6 +27,7 @@ import androidx.fragment.app.FragmentActivity;
 import androidx.lifecycle.Lifecycle;
 import io.flutter.Log;
 import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.embedding.engine.FlutterShellArgs;
 import io.flutter.embedding.engine.renderer.FlutterUiDisplayListener;
 import io.flutter.plugin.platform.PlatformPlugin;
 import io.flutter.plugin.view.SensitiveContentPlugin;
@@ -255,7 +256,7 @@ public class FlutterFragment extends Fragment
     private String initialRoute = "/";
     private boolean handleDeeplinking = false;
     private String appBundlePath = null;
-    private String[] shellArgs = null;
+    private FlutterShellArgs shellArgs = null;
     private RenderMode renderMode = RenderMode.surface;
     private TransparencyMode transparencyMode = TransparencyMode.transparent;
     private boolean shouldAttachEngineToActivity = true;
@@ -331,7 +332,7 @@ public class FlutterFragment extends Fragment
 
     /** Any special configuration arguments for the Flutter engine */
     @NonNull
-    public NewEngineFragmentBuilder flutterShellArgs(@NonNull String[] shellArgs) {
+    public NewEngineFragmentBuilder flutterShellArgs(@NonNull FlutterShellArgs shellArgs) {
       this.shellArgs = shellArgs;
       return this;
     }
@@ -458,8 +459,9 @@ public class FlutterFragment extends Fragment
           dartEntrypointArgs != null ? new ArrayList(dartEntrypointArgs) : null);
       // TODO(mattcarroll): determine if we should have an explicit FlutterTestFragment instead of
       // conflating.
-      args.putStringArray(
-          ARG_FLUTTER_INITIALIZATION_ARGS, shellArgs == null ? new String[0] : shellArgs);
+      if (null != shellArgs) {
+        args.putStringArray(ARG_FLUTTER_INITIALIZATION_ARGS, shellArgs.toArray());
+      }
       args.putString(
           ARG_FLUTTERVIEW_RENDER_MODE,
           renderMode != null ? renderMode.name() : RenderMode.surface.name());
@@ -1351,9 +1353,10 @@ public class FlutterFragment extends Fragment
    */
   @Override
   @NonNull
-  public String[] getFlutterShellArgs() {
+  public FlutterShellArgs getFlutterShellArgs() {
     String[] flutterShellArgsArray = getArguments().getStringArray(ARG_FLUTTER_INITIALIZATION_ARGS);
-    return flutterShellArgsArray == null ? new String[0] : flutterShellArgsArray;
+    return new FlutterShellArgs(
+        flutterShellArgsArray != null ? flutterShellArgsArray : new String[] {});
   }
 
   /**
