+++ engine/src/flutter/shell/platform/android/io/flutter/embedding/engine/FlutterEngine.java
@@ -177,7 +177,7 @@ public class FlutterEngine implements ViewUtils.DisplayUpdater {
    * native library and start a Dart VM.
    *
    * <p>In order to pass Dart VM initialization arguments (see {@link
-   * io.flutter.embedding.engine.FlutterShellArgs} for all available flags) when creating the VM,
+   * io.flutter.embedding.engine.FlutterEngineFlags} for all available flags) when creating the VM,
    * manually set the initialization arguments by calling {@link
    * io.flutter.embedding.engine.loader.FlutterLoader#startInitialization(Context)} and {@link
    * io.flutter.embedding.engine.loader.FlutterLoader#ensureInitializationComplete(Context,
